requirements.txt
aiogram==3.17.0
crewai==0.95.0
crewai-tools==0.25.8
pydantic==2.10.5
python-dotenv==1.0.1
fastapi==0.109.1
uvicorn==0.27.0
hivetrace==1.2.5
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.12.0
